backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-socketio==5.5.0
python-dotenv==0.19.0
google-cloud-speech==2.11.0
google-cloud-translate==3.7.0
google-cloud-texttospeech==2.9.0
motor==2.5.1
pymongo==3.12.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.5 